requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
pytest-html==4.1.1
pytest-xdist==3.3.1
coverage==7.3.2
flake8==6.1.0
black==23.11.0 